requirements.txt
langchain
langgraph
langgraph-cli
langgraph-api 
markdown    
weasyprint 
beautifulsoup4==4.14.2
langchain_community==0.4.1
langchain_core==1.0.5
langchain_text_splitters==1.0.0
pandas==2.3.3
qdrant_client==1.16.0
Requests==2.32.5
sentence_transformers==5.1.2
tqdm==4.67.1
sentence_transformers
google-genai
langchain-google-genai
streamlit

